+++ flutter/lib/src/sentry_widget.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:meta/meta.dart';
 import '../sentry_flutter.dart';
+import 'utils/multi_view_helper.dart';
 
 /// Key which is used to identify the [SentryWidget]
 @internal
@@ -11,7 +12,13 @@ final sentryWidgetGlobalKey = GlobalKey(debugLabel: 'sentry_widget');
 class SentryWidget extends StatefulWidget {
   final Widget child;
 
-  const SentryWidget({super.key, required this.child});
+  SentryWidget({
+    super.key,
+    required this.child,
+    @internal Hub? hub,
+  });
+
+  final bool _isMultiViewEnabled = MultiViewHelper.isMultiViewEnabled();
 
   @override
   _SentryWidgetState createState() => _SentryWidgetState();
@@ -21,6 +28,14 @@ class _SentryWidgetState extends State<SentryWidget> {
   @override
   Widget build(BuildContext context) {
     Widget content = widget.child;
+    if (widget._isMultiViewEnabled) {
+      // ignore: invalid_use_of_internal_member
+      Sentry.currentHub.options.logger(
+        SentryLevel.debug,
+        '`SentryScreenshotWidget` and `SentryUserInteractionWidget` is not available in multi-view applications.',
+      );
+      return content;
+    }
     content = SentryScreenshotWidget(child: content);
     content = SentryUserInteractionWidget(child: content);
     return Container(