+++ flutter/lib/src/utils/multi_view_helper.dart
@@ -1,0 +1,14 @@
+import 'dart:ui';
+import 'package:meta/meta.dart';
+
+@internal
+class MultiViewHelper {
+  static bool isMultiViewEnabled() {
+    final dynamic uncheckedImplicitView = PlatformDispatcher.instance;
+    try {
+      return null == uncheckedImplicitView.implicitView;
+    } on NoSuchMethodError catch (_) {
+      return false;
+    }
+  }
+}